refactor for python package
