+++ autotune/cache/parameter_importance.py
@@ -3,7 +3,6 @@ import json
 import os
 
 import matplotlib.pyplot as plt
-import numpy as np
 import pandas as pd
 import seaborn as sns
